Line items can no longer be deleted if they have fulfillments. [#76889628]

+++ app/assets/javascripts/portal/fulfillment2.js.coffee
@@ -110,11 +110,24 @@ $(document).ready ->
   $(document).on('click', '.delete_data', ->
     klass = getObjKlass(this)
     object_id = $(this).data("#{klass}_id")
-    data = {}
-    data['study_tracker'] = $('#study_tracker_hidden_field').val() || null
-    confirm_message = "Are you sure that you want to remove this service from all subjects' visit calendars in this arm?"
-    if $(this).data("has_popup") == true
-      if confirm(confirm_message)
+    has_fulfillments = $(this).data("has_fulfillments") || null
+    if has_fulfillments
+      alert("This line item has one or more fulfillments associated with it and can not be deleted.")
+    else
+      data = {}
+      data['study_tracker'] = $('#study_tracker_hidden_field').val() || null
+      confirm_message = "Are you sure that you want to remove this service from all subjects' visit calendars in this arm?"
+      if $(this).data("has_popup") == true
+        if confirm(confirm_message)
+          $.ajax
+            type: 'DELETE'
+            url:  "/portal/admin/#{klass}s/#{object_id}"
+            data: JSON.stringify(data)
+            dataType: "script"
+            contentType: 'application/json; charset=utf-8'
+            success: ->
+              $().toastmessage('showSuccessToast', "#{klass.humanize()} has been deleted.");
+      else
         $.ajax
           type: 'DELETE'
           url:  "/portal/admin/#{klass}s/#{object_id}"
@@ -123,15 +136,6 @@ $(document).ready ->
           contentType: 'application/json; charset=utf-8'
           success: ->
             $().toastmessage('showSuccessToast', "#{klass.humanize()} has been deleted.");
-    else
-      $.ajax
-        type: 'DELETE'
-        url:  "/portal/admin/#{klass}s/#{object_id}"
-        data: JSON.stringify(data)
-        dataType: "script"
-        contentType: 'application/json; charset=utf-8'
-        success: ->
-          $().toastmessage('showSuccessToast', "#{klass.humanize()} has been deleted.");
   )
 
   $('#cwf_building_dialog').dialog
@@ -444,18 +448,22 @@ $(document).ready ->
   $(document).on('click', '.cwf_delete_data', ->
     klass = getObjKlass(this)
     object_id = $(this).data("#{klass}_id")
-    data = {}
-    data['study_tracker'] = $('#study_tracker_hidden_field').val() || null
-    confirm_message = "Are you sure that you want to remove this service?"
-    if confirm(confirm_message)
-      $.ajax
-        type: 'DELETE'
-        url:  "/portal/admin/line_items/#{object_id}"
-        data: JSON.stringify(data)
-        dataType: "script"
-        contentType: 'application/json; charset=utf-8'
-        success: ->
-          $().toastmessage('showSuccessToast', "Service has been deleted.");
+    has_fulfillments = $(this).data("has_fulfillments") || null
+    if has_fulfillments
+      alert("This line item has one or more fulfillments associated with it and can not be deleted.")
+    else
+      data = {}
+      data['study_tracker'] = $('#study_tracker_hidden_field').val() || null
+      confirm_message = "Are you sure that you want to remove this service?"
+      if confirm(confirm_message)
+        $.ajax
+          type: 'DELETE'
+          url:  "/portal/admin/line_items/#{object_id}"
+          data: JSON.stringify(data)
+          dataType: "script"
+          contentType: 'application/json; charset=utf-8'
+          success: ->
+            $().toastmessage('showSuccessToast', "Service has been deleted.");
   )
 
   $(document).on('click', '.expand_li', ->
